load_test_data/requirements.txt
psycopg-binary==3.2.9
psycopg==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
